feat: 음소거 버튼 기능 구현
- 각 스테이지 별 배경음악 설정에 필요한 기본 세팅을 다른 컴포넌트로 분리
- 음악 버튼 클릭 시 배경화면 음소거 또는 재생 기능 구현
- 스테이지별 백그라운드 설정
- 게임 메뉴 컴포넌트 재 구성
