+++ src/components/menus/GameMenu.js
@@ -2,7 +2,6 @@ import styled, { css } from "styled-components";
 import React, { useState } from "react";
 
 import { ReactComponent as Reload } from "../../assets/icon/Reload.svg";
-import { ReactComponent as Sound } from "../../assets/icon/Sound.svg";
 import { ReactComponent as StageSelection } from "../../assets/icon/StageSelection.svg";
 import Modal from "../modals/Modal";
 
@@ -29,7 +28,6 @@ export default function GameMenu() {
       )}
       <ResetButton onClick={handleResetButtonClick} />
       <StageSelectionButton onClick={handleStageSelectButtonClick} />
-      <SoundControlButton />
     </>
   );
 }
@@ -44,12 +42,6 @@ const commonStyles = css`
 const ResetButton = styled(Reload)`
   ${commonStyles};
   right: 13vw;
-`;
-
-const SoundControlButton = styled(Sound)`
-  ${commonStyles};
-  right: 2vw;
-  top: 7.5vh;
 `;
 
 const StageSelectionButton = styled(StageSelection)`
