+++ src/components/stages/StageThree.js
@@ -5,10 +5,12 @@ import { OrbitControls } from "@react-three/drei";
 import stageThreeCoordinates from "../../data/stageThreeCoordinates.json";
 import Cube from "../objects/Cube";
 import GameMenu from "../menus/GameMenu";
+import BackgroundMusicSetting from "../music/BackgroundMusicSetting";
 
 export default function StageThree() {
   return (
     <>
+      <BackgroundMusicSetting />
       <Canvas
         camera={{
           position: [15, 15, 15],
